Fix infinite loop on MacOS
We're getting numerical differences on MacOS vs other platforms, and this can cause an infinite loop in parts of the library.  Using std::fma seems to fix the differences.  I fought clang for a while to make it define the FP_FAST_FMA* macros, but I failed.  As noted in the comment, I confirmed in a Godbolt session that passing -mfma would still use the intrinsic version.

```
float func(float a, float b) {
	return std::fmaf(a, b, 3);
}

#ifndef FP_FAST_FMAF
#error "Fast FMA operation not supported"
#endif
```

Compiling with `clang -mfma` will trigger the `#error` statement but will produce a `vfmadd213ss` call if you remove the `FP_FAST_FMAF` check.

This commit leaves the code in a weird spot, but I'm avoiding a deeper change so we don't run into issue pulling in any new upstream changes.